refactor: useExpression hook 정의
